Update consent only on user choise

+++ src/tools/gtag.ts
@@ -57,25 +57,27 @@ function deleteGaCookies() {
 }
 
 export function setConsentSettings(allowConsent?: boolean) {
-    let analyticsConsent = hasUserConsent(allowConsent) ? "granted" : "denied";
+    if (!!allowConsent) {
+        let analyticsConsent = hasUserConsent(allowConsent) ? "granted" : "denied";
 
-    if (analyticsConsent === "denied") {
-        deleteGaCookies();
+        if (analyticsConsent === "denied") {
+            deleteGaCookies();
 
-        if (data.gtag) {
-            window[`ga-disable-${data.gtag}`] = true;
+            if (data.gtag) {
+                window[`ga-disable-${data.gtag}`] = true;
+            }
+            window[`ga-disable-${ga_common_prop}`] = true;
+        } else {
+            if (data.gtag) {
+                window[`ga-disable-${data.gtag}`] = false;
+            }
+            window[`ga-disable-${ga_common_prop}`] = false;
         }
-        window[`ga-disable-${ga_common_prop}`] = true;
-    } else {
-        if (data.gtag) {
-            window[`ga-disable-${data.gtag}`] = false;
-        }
-        window[`ga-disable-${ga_common_prop}`] = false;
+
+        gtag("consent", "update", {
+            analytics_storage: analyticsConsent,
+        });
     }
-
-    gtag("consent", "update", {
-        analytics_storage: analyticsConsent,
-    });
 }
 
 export function sendEventToGa(action: GaEventActions, label: string, eventCategory?: string) {
